requirements.txt
Flask
yt_dlp
youtube-search-python
requests
python-dotenv
gunicorn

https://rr5---sn-5ucgv5qc5oq-ug8e.googlevideo.com/videoplayback?expire=1734467494&ei=RothZ7H_HcSrssUPpe3CwAE&ip=103.169.56.47&id=o-AOjlgw5mgG8_035jK1V19c1xK7nVbG1nmXmavEeWc7CN&itag=251&source=youtube&requiressl=yes&xpc=EgVo2aDSNQ%3D%3D&met=1734445894%2C&mh=Lp&mm=31%2C29&mn=sn-5ucgv5qc5oq-ug8e%2Csn-h5576nss&ms=au%2Crdu&mv=m&mvi=5&pl=24&rms=au%2Cau&initcwndbps=2271250&bui=AfMhrI9PrHzrUD0aYx_HRONCisgiP6fbutkSfK9ji61yfO2i26xbqg9AsPAWu_7H5h07g05Jfq-7Zktn&spc=x-caUNsLwiKibsPku4gcfzhiyA5HCcQGtNF1bvlbinAOANY99g&vprv=1&svpuc=1&mime=audio%2Fwebm&rqh=1&gir=yes&clen=6498427&dur=376.661&lmt=1728318578351308&mt=1734445530&fvip=4&keepalive=yes&fexp=51326932%2C51335594&c=IOS&txp=4532434&sparams=expire%2Cei%2Cip%2Cid%2Citag%2Csource%2Crequiressl%2Cxpc%2Cbui%2Cspc%2Cvprv%2Csvpuc%2Cmime%2Crqh%2Cgir%2Cclen%2Cdur%2Clmt&sig=AJfQdSswRgIhAPhOOdI3TVSOVh_zUJvd2rP_zyFtrg8QDgxYsj6DQGpzAiEAlWpP1_6xavpQu5hhh2bIW0fuITESL8mjHDW9ZzqrvBA%3D&lsparams=met%2Cmh%2Cmm%2Cmn%2Cms%2Cmv%2Cmvi%2Cpl%2Crms%2Cinitcwndbps&lsig=AGluJ3MwRAIgGirRZw6268ZiZUakz3mPz176THRXXsEMQ1ipO8bx0nQCIGja38G7FCF2A5VKc3pqUIBYD5KJOZlSNE-urQ4s9KIi